pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0", 
    "wheel", 
    "setuptools_scm[toml]>=6.2",
]
build-backend = "setuptools.build_meta"

[project]
name = "pyvvisf"
dynamic = ["version"]
description = "Pure Python ISF shader renderer with PyOpenGL and json5"
readme = "README.md"
license = "MIT"
license-files = ["LICENSE"]
authors = [
    {name = "Jim Cortez", email = "gitstuff@jimcortez.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "glfw>=2.5.0",
    "PyOpenGL>=3.1.0",
    "PyOpenGL-accelerate>=3.1.0",
    "json5>=0.9.0",
    "pillow>=9.0.0",
    "numpy>=1.21.0",
    "pydantic>=1.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
    "pre-commit>=2.20.0",
    "cibuildwheel>=3.0.0",
    "build>=0.10.0",
]

[project.urls]
Homepage = "https://github.com/jimcortez/pyvvisf"
Repository = "https://github.com/jimcortez/pyvvisf"
Documentation = "https://github.com/jimcortez/pyvvisf#readme"
Issues = "https://github.com/jimcortez/pyvvisf/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools_scm]
write_to = "src/pyvvisf/_version.py"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
] 